// src/lib/api/gen/types/ResubmitExpenseDto.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

export type ResubmitExpenseDto = {
  /**
   * @type string | undefined
   */
  title?: string;
  /**
   * @type object | undefined
   */
  description?: object;
  /**
   * @minLength 0
   * @type number | undefined
   */
  amount?: number;
  /**
   * @default "BDT"
   * @type string | undefined
   */
  currency?: string;
  /**
   * @type string | undefined
   */
  category?: string;
  /**
   * @type object | undefined
   */
  vendor?: object;
  /**
   * @type object | undefined
   */
  itemValue?: object;
  /**
   * @type object | undefined
   */
  price?: object;
  /**
   * @type object | undefined
   */
  quantity?: object;
  /**
   * @type object | undefined
   */
  departmentId?: object;
  /**
   * @type object | undefined
   */
  customFieldsJson?: object;
};